# packages/api/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/admin/tenants": {
      "get": {
        "operationId": "TenantsController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "platform-auth": []
          }
        ],
        "summary": "List all tenants",
        "tags": [
          "Admin - Tenants"
        ]
      },
      "post": {
        "operationId": "TenantsController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateTenantDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "platform-auth": []
          }
        ],
        "summary": "Create a new tenant",
        "tags": [
          "Admin - Tenants"
        ]
      }
    },
    "/admin/tenants/{id}": {
      "get": {
        "operationId": "TenantsController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "platform-auth": []
          }
        ],
        "summary": "Get tenant by ID",
        "tags": [
          "Admin - Tenants"
        ]
      }
    },
    "/admin/tenants/{id}/suspend": {
      "patch": {
        "operationId": "TenantsController_suspend",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "platform-auth": []
          }
        ],
        "summary": "Suspend a tenant",
        "tags": [
          "Admin - Tenants"
        ]
      }
    },
    "/admin/tenants/{id}/archive": {
      "patch": {
        "operationId": "TenantsController_archive",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "platform-auth": []
          }
        ],
        "summary": "Archive a tenant",
        "tags": [
          "Admin - Tenants"
        ]
      }
    },
...
    "/sites": {
      "get": {
        "operationId": "SitesController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "tenant-auth": []
          }
        ],
        "summary": "List sites",
        "tags": [
          "Sites"
        ]
      },
      "post": {
        "operationId": "SitesController_create",
        "parameters": [],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "tenant-auth": []
          }
        ],
        "summary": "Create a site",
        "tags": [
          "Sites"
        ]
      }
    },
    "/sites/{id}": {
      "get": {
        "operationId": "SitesController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "tenant-auth": []
          }
        ],
        "summary": "Get site by ID",
        "tags": [
          "Sites"
        ]
      }
    },
...
    "/photos/presign": {
      "post": {
        "operationId": "PhotosController_presign",
        "parameters": [],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "tenant-auth": []
          }
        ],
        "summary": "Get an upload URL/path",
        "tags": [
          "Photos"
        ]
      }
    },
    "/photos/{id}/upload": {
      "post": {
        "operationId": "PhotosController_upload",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "file": {
                    "type": "string",
                    "format": "binary"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "tenant-auth": []
          }
        ],
        "summary": "Upload a photo file",
        "tags": [
          "Photos"
        ]
      }
    },
    "/photos/{id}/commit": {
      "post": {
        "operationId": "PhotosController_commit",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "tenant-auth": []
          }
        ],
        "summary": "Commit an uploaded photo with SHA-256",
        "tags": [
          "Photos"
        ]
      }
    },
    "/photos": {
      "get": {
        "operationId": "PhotosController_getPhotos",
        "parameters": [
          {
            "name": "entity_type",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "entity_id",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "tenant-auth": []
          }
        ],
        "summary": "List photos for an entity",
        "tags": [
          "Photos"
        ]
      }
    },
    "/photos/{id}": {
      "get": {
        "operationId": "PhotosController_getPhoto",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "tenant-auth": []
          }
        ],
        "summary": "Download a photo",
        "tags": [
          "Photos"
        ]
      }
    },
    "/photos/{id}/url": {
      "get": {
        "operationId": "PhotosController_getUrl",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "tenant-auth": []
          }
        ],
        "summary": "Get photo download URL",
        "tags": [
          "Photos"
        ]
      }
    },
...
    "/reports/projection-inputs": {
      "get": {
        "operationId": "ReportsController_getInputs",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "tenant-auth": []
          }
        ],
        "summary": "Projection defaults for this tenant (RPT-05)",
        "tags": [
          "Reports"
        ]
      }
    },
    "/reports/projections": {
      "get": {
        "operationId": "ReportsController_project",
        "parameters": [
          {
            "name": "working_days",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "units_per_day",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "rate_minor",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "currency",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "tenant-auth": []
          }
        ],
        "summary": "Project billing and contribution from working days × units × rate",
        "tags": [
          "Reports"
        ]
      }
    },
...
  },
  "info": {
    "title": "Fleet OS API",
    "description": "Multi-tenant SaaS for heavy-equipment operators",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "schemas": {
      "CreateTenantDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "Acme Equipment"
          },
          "slug": {
            "type": "string",
            "example": "acme"
          },
          "country": {
            "type": "string",
            "example": "NG"
          },
          "base_currency": {
            "type": "string",
            "example": "NGN"
          },
          "timezone": {
            "type": "string",
            "example": "Africa/Lagos"
          }
        },
        "required": [
          "name",
          "slug",
          "country",
          "base_currency"
        ]
      },
...
    }
  }
}